+++ pattern_recognition/src/parallelManualAlgoNoFirstTouch.cpp
@@ -8,7 +8,7 @@ ParallelManualAlgoNoFirstTouch::ParallelManualAlgoNoFirstTouch()
     this->vectorized = true;
 }
 
-void ParallelManualAlgoNoFirstTouch::initData(Data* data, int size, float* values) {
+void ParallelManualAlgoNoFirstTouch::initData(Data* data, long long int size, float* values) {
     data->size = size;
     data->values = new float[size];
     for(int i = 0; i < size; i++) {
